+++ templates/website/design.js
@@ -74,6 +74,7 @@ module.exports = cx.design
         require('@bsi-cx/design-standard-library-web/content-elements/form/advanced-file-upload'),
         require("@bsi-cx/design-standard-library-web/content-elements/form/text-area"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/select"),
+        require("@bsi-cx/design-standard-library-web/content-elements/form/dynamic-select"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/checkbox"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/radio"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/poll"),
@@ -82,6 +83,7 @@ module.exports = cx.design
         // require("@bsi-cx/design-standard-library-web/content-elements/form/regex-input"), !!! set properties.regexPattern first !!!
       )
   )
+  .withFeatureFormFieldRules(true)
   .withWebsite(
     cx.website
       .withMaxNavigationLevel(2)