+++ CHANGELOG.md
@@ -8,7 +8,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Hint: You will find the list of compatible BSI versions for each template in the [Compatibility section of the README](https://github.com/bsi-software/bsi-cx-design-master-template-web?tab=readme-ov-file#compatibility).
 
 ## Next
-
+* Add Feature FormfieldRules
+* Add dynamic select from standard library
 
 
 ## [1.3.10] - 15.04.2026